+++ peepholelib/coreVectors/activations.py
@@ -21,70 +21,57 @@ def get_activations(self, **kwargs):
     bs = kwargs['batch_size'] if 'batch_size' in kwargs else 64
     n_threads = kwargs['n_threads'] if 'n_threads' in kwargs else 1 
     verbose = kwargs['verbose'] if 'verbose' in kwargs else False
+    save_input = kwargs['save_input'] if 'save_input' in kwargs else True
+    save_output = kwargs['save_output'] if 'save_output' in kwargs else False 
 
     model = self._model
-    num_classes = self._model.num_classes
     device = self._model.device 
-    hooks = model.get_hooks()
     
     assert(isinstance(model, ModelWrap))
     if self._dss == {}:
         raise RuntimeError('No dataset parsed. Run `parse_ds()` first.')
     
+    # set the model to get activations
+    model.set_activations(save_input=save_input, save_output=save_output)
+
     for ds_key in self._dss:
         if verbose: print(f'\n ---- Getting data from {ds_key}\n')
+        n_samples = len(self._dss[ds_key])
        
         #------------------------------------------------
         # pre-allocate activations
         #------------------------------------------------
 
-        # check if in and out activations exist
-        if model._si:
-            if not ('in_activations' in self._dss[ds_key]):
-                if verbose: print('adding in act tensorDict')
-                self._dss[ds_key]['in_activations'] = TensorDict(batch_size=n_samples)
-            else: 
-                if verbose: print('In activations exist.')
-        if 'in_activations' in self._dss[ds_key]: self._dss[ds_key]['in_activations'].batch_size = torch.Size((n_samples,)) 
-
-        if model._so:
-            if not ('out_activations' in self._dss[ds_key]):
-                if verbose: print('adding out act tensorDict')
-                self._dss[ds_key]['out_activations'] = TensorDict(batch_size=n_samples)
-            else: 
-                if verbose: print('Out activations exist.')
-        if 'out_activations' in self._dss[ds_key]: self._dss[ds_key]['out_activations'].batch_size = torch.Size((n_samples,)) 
-        
         # dry run to get shapes
-        img = self._dss[ds_key][0:1]['image'].to(device)
-        print('img shape: ', img.shape)
-        model(img)
+        with torch.no_grad():
+            model(self._dss[ds_key]['image'][0:1].to(device))
 
-        # check if module exists in in_ and out_activations
-        _modules_to_save = []
-        for mk in model.get_target_modules():
-            # prevents double entries 
-            _lts = None
-
-            # allocate for input activations 
-            if model._si and (not (mk in self._dss[ds_key]['in_activations'])):
-                if verbose: print('allocating in act module: ', mk)
-                # Seems like when loading from memory the batch size gets overwritten with all dims, so we over-overwrite it.
-                act_shape = hooks[mk].in_shape
-                self._dss[ds_key]['in_activations'][mk] = MMT.empty(shape=torch.Size((n_samples,)+act_shape))
-                _lts = mk
-
-            # allocate for output activations 
-            if model._so and (not (mk in self._dss[ds_key]['out_activations'])):
-                if verbose: print('allocating out act module: ', mk)
-                act_shape = hooks[mk].out_shape
-                self._dss[ds_key]['out_activations'][mk] = MMT.empty(shape=torch.Size((n_samples,)+act_shape))
-                _lts = mk
+        _modules_to_save = {}
+        for act_key in model._acts:
+            if not (act_key in self._dss[ds_key]):
+                if verbose: print(f'adding {act_key} tensorDict')
+                self._dss[ds_key][act_key] = TensorDict(batch_size=n_samples)
+            else:
+                # this is a workaound a pesky bug when loading PTDs with thensors on it, which sets bs = data.shape instead of data.shape[0]
+                self._dss[ds_key][act_key].batch_size = torch.Size((n_samples,)) 
             
-            if _lts != None: _modules_to_save.append(_lts)
-        
+            _modules_to_save[act_key] = []
+            for mk in model._acts[act_key]:
+                # allocate activations 
+                if not (mk in self._dss[ds_key][act_key]):
+                    if verbose: print(f'allocating {act_key} for {mk}')
+                    act_shape = model._acts[act_key][mk].shape[1:]
+                    self._dss[ds_key][act_key][mk] = MMT.empty(shape=torch.Size((n_samples,)+act_shape))
+                    _modules_to_save[act_key].append(mk)
+
         if verbose: print('modules to save: ', _modules_to_save)
-        if len(_modules_to_save) == 0:
+        # check if there is anything to save for ds_key
+        skip = True
+        for ms in _modules_to_save.values():
+            if len(ms) != 0:
+               skip = False 
+        
+        if skip:
             if verbose: print(f'No new activations for {ds_key}, skipping')
             continue
 
@@ -98,13 +85,14 @@ def get_activations(self, **kwargs):
         if verbose: print('Computing activations')
         for data in tqdm(ds_dl, disable=not verbose):
             with torch.no_grad():
-                model(ds_data['image'].to(device))
+                model(data['image'].to(device))
             
-            for mk in _modules_to_save:
-                if model._si:
-                    data['in_activations'][mk] = hooks[mk].in_activations[:].cpu()
+            for act_key in _modules_to_save:
+                for mk in _modules_to_save[act_key]:
+                    data[act_key][mk] = model._acts[act_key][mk].cpu()
 
-                if model._so:
-                    data['out_activations'][mk] = hooks[mk].out_activations[:].cpu()
+    # reset the model to NOT get activations
+    model.set_activations(save_input=False, save_output=False)
+
     return