+++ RELEASENOTES.md
@@ -6,7 +6,8 @@
 ## Model
 - add function `update_output` to `model_wrap`
 - Remove dry_run(). Corevectors run a dry image internally.
-- Revine `add_hooks()`, not it is done in `set_target_modules()`
+- Remove `add_hooks()`, not it is done in `set_target_modules()`
+- Add `set_activations()` function, which set the model to save activation in `model._acts` or to not save activations.
 
 ## SVDs
 - Support `torch.nn.Conv2d` layers with no bias, and with groups